+++ routes/web.php
@@ -19,18 +19,26 @@ use App\Http\Controllers\ListingController;
 // All listing
 Route::get('/', [ListingController::class, 'index']);
 // post creation
-Route::get('/listings/create', [ListingController::class, 'create']);
+Route::get('/listings/create', [ListingController::class, 'create'])->middleware('auth');
 // store listing
-Route::post('/listings', [ListingController::class, 'store']);
+Route::post('/listings', [ListingController::class, 'store'])->middleware('auth');
+// manage listing
+Route::get('/listings/manage', [ListingController::class, 'manage'])->middleware('auth');
 // single listing
 Route::get('/listings/{listing}', [ListingController::class, 'show']);
 // update listing
-Route::put('/listings/{listing}', [ListingController::class, 'update']);
+Route::put('/listings/{listing}', [ListingController::class, 'update'])->middleware('auth');
 // delete listing
-Route::delete('/listings/{listing}', [ListingController::class, 'destroy']);
+Route::delete('/listings/{listing}', [ListingController::class, 'destroy'])->middleware('auth');
 // edit listing
-Route::get('/listings/{listing}/edit', [ListingController::class, 'edit']);
+Route::get('/listings/{listing}/edit', [ListingController::class, 'edit'])->middleware('auth');
 // register form
-Route::get('/register', [UserController::class, 'create']);
+Route::get('/register', [UserController::class, 'create'])->middleware('guest');
 // create new user
 Route::post('/users', [UserController::class, 'store']);
+// Logout
+Route::post('/logout', [UserController::class, 'logout'])->middleware('auth');
+// show login form
+Route::get('/login', [UserController::class, 'login'])->name('login')->middleware('guest');
+// user login
+Route::post('/users/authenticate', [UserController::class, 'authenticate']);
